rename to ncep_obs
