+++ earth2studio/data/ncep_obs.py
